fix: updated patch, add fields only if fields are checked
